{"schema":"decodex/commit/1","summary":"Modularize macOS window bridge","authority":"manual"}
